feat: 관리자 대시보드 차트/통계 연동, 기부 관리 페이지, Featured 메시지, 자동 로그아웃, UI 디자인 통일

+++ backend/src/main/java/com/wenect/donation_paltform/domain/admin/dto/DashboardStatsResponse.java
@@ -25,11 +25,11 @@ public class DashboardStatsResponse {
     // 지난주 대비 신규 회원 증감률 (%)
     private Double userChange;
 
-    // 승인 대기 (목업 데이터)
-    private Integer pendingApprovals;
+    // 기관 승인 대기
+    private Long pendingApprovals;
 
-    // 정산 승인 대기 (목업 데이터)
-    private Integer pendingSettlements;
+    // 정산 승인 대기
+    private Long pendingSettlements;
 
     // 지출 승인 대기
     private Long pendingExpenses;
